refactor(vm): Move the out_of_memory test to tests/limits.rs
